# Use faster serializers for Search::Postgres::Tag