--- a/app/services/search/postgres/tag.rb
+++ b/app/services/search/postgres/tag.rb
@@ -4,13 +4,14 @@ module Search
       ATTRIBUTES = %i[id name hotness_score rules_html supported short_summary].freeze
 
       def self.search_documents(term)
-        ::Tag
-          .search_by_name(term)
-          .where(supported: true)
-          .reorder(hotness_score: :desc)
-          .select(*ATTRIBUTES)
-          .as_json
+        results = ::Tag.search_by_name(term).where(supported: true).reorder(hotness_score: :desc).select(*ATTRIBUTES)
+        serialize(results)
       end
+
+      def self.serialize(results)
+        Search::TagSerializer.new(results, is_collection: true).serializable_hash[:data].pluck(:attributes)
+      end
+      private_class_method :serialize
     end
   end
 end